Change dependency to only count outbound violations and privacy to only count inbound violations
It was a bit confusing that previously, by_package.dependency violations
counted a package's outbound AND inbound violations. This creates double
counting of violations AND we should only care about outbound dependency
violations anyways.

`by_package.outbound_privacy_violations.per_package` was also a bit of a
misnomer, because we were actually showing the results for INBOUND
violations into the selected package. This was mostly an ergonomic thing
(because this was the view of per package we wanted, and we didn't want
folks to have to select their pack in "to_package."

Instead, we should rename this
by_package.privacy_violations.per_package, which shows the inbound
privacy violations for the selected package, and we should rename
to_package to "other_package"

to_team => other_team and to_package => other_package

The idea here is that this template var is no longer used to represent
"directionality" of a violation. Instead, it's used to represent the
other team or package involved in a violation we care about.

For example, with by_package.privacy_violations.per_package, the
"other_package" is the package that is doing the violating.

However for by_package.dependency_violations.per_package, the
"other_package" is the package being violated.

We actually don't care about the directionality in this context.

We also actually don't care about who the other is. That is, it would
never make sense to view the pack stats dashboard and look at
"other_package" or "other_team." They're only to be used to display a
widget of violation pairs, but we can actually remove them from the
stats dashboard entirely!