per_x => by_other_x
I didn't like the use of the preposition "per" in "per_package" and
"per_team."

The fact that it was something other than "by" (which was the previous
part of the metric, i.e. by_package, by_team) suggested there was
something different about it.

In reality, we are just grouping by package and by other package. Using
the same preposition made it clearer to me (and hopefully the reader)
what the metric was for, and the fact that it matched up to the tag name
makes extra sense.