Revamp of metrics (#35)
* Remove inbound_dependency_violations

We do not care about inbound dependency violations – only outbound.

* outbound_dependency_violations => dependency_violations

When we say "dependency violations," we always mean outbound. This makes
it more concise and removes the unnecessary and confusing term
"outbound."

* inbound_privacy_violations => privacy_violations

When we say privacy violations, we always mean "inbound."

The distinction is confusing since our area of focus should always be
inbound privacy violations.

* stop tracking outbound_privacy_violations

Outbound privacy violations are not particularly interesting. While it
may be a good idea to solve, the area of focus should always be inbound
privacy violations first.

* Change dependency to only count outbound violations and privacy to only count inbound violations

It was a bit confusing that previously, by_package.dependency violations
counted a package's outbound AND inbound violations. This creates double
counting of violations AND we should only care about outbound dependency
violations anyways.

`by_package.outbound_privacy_violations.per_package` was also a bit of a
misnomer, because we were actually showing the results for INBOUND
violations into the selected package. This was mostly an ergonomic thing
(because this was the view of per package we wanted, and we didn't want
folks to have to select their pack in "to_package."

Instead, we should rename this
by_package.privacy_violations.per_package, which shows the inbound
privacy violations for the selected package, and we should rename
to_package to "other_package"

to_team => other_team and to_package => other_package

The idea here is that this template var is no longer used to represent
"directionality" of a violation. Instead, it's used to represent the
other team or package involved in a violation we care about.

For example, with by_package.privacy_violations.per_package, the
"other_package" is the package that is doing the violating.

However for by_package.dependency_violations.per_package, the
"other_package" is the package being violated.

We actually don't care about the directionality in this context.

We also actually don't care about who the other is. That is, it would
never make sense to view the pack stats dashboard and look at
"other_package" or "other_team." They're only to be used to display a
widget of violation pairs, but we can actually remove them from the
stats dashboard entirely!

* Get rid of "enforcing_x" metrics

We already ahve this in the form
all_packages.packwerk_checkers.enforce_x.strict|true|false.count

We can always sum the maximum of these values to get count of all true
or strict.

* Get rid of with_violations metric

We don't use this, and it's superceded by other metrics that count total
number of violations. We can always have a DD query that checks for
violations = 0.

* outbound_explicit_dependencies => dependencies

We can simply call these "dependencies" of a package to be concise, as
that is what is in the package.yml

* inbound_explicit_dependencies => depended_on

I think "depended on" is clearer than "inbound explicit dependencies"

* Remove use_gusto_legacy_names

We are going to move to using the currently supported public API.

* update README

* sort lines

* factor out dependencies metrics into own class

* Move variable closer to use

* add packwerk checker usage

* version

* fix bug in by_team.violations.per_team

I had a bug in the way that I got the team-to-team violations.

I fixed the bug by clarifying the variable names within the class and
then fixing what violations were selected.

* per_x => by_other_x

I didn't like the use of the preposition "per" in "per_package" and
"per_team."

The fact that it was something other than "by" (which was the previous
part of the metric, i.e. by_package, by_team) suggested there was
something different about it.

In reality, we are just grouping by package and by other package. Using
the same preposition made it clearer to me (and hopefully the reader)
what the metric was for, and the fact that it matched up to the tag name
makes extra sense.

* simplify code to get checker metrics

* fix bug with packwerk checker usage

* fix bug in by_package.violations.by_other_package

This is similar to the same bug detected in
by_team.violations.by_other_team